refactor: move files related to SingleStat
