+++ giraffe/src/components/SingleStat/SingleStatLayer.tsx
@@ -6,9 +6,9 @@ import {
   SINGLE_STAT_SVG_DEFAULT_ATTRIBUTES,
   SINGLE_STAT_SVG_TEXT_DEFAULT_ATTRIBUTES,
   SINGLE_STAT_SVG_TEXT_DEFAULT_STYLE,
-} from '../constants/singleStatStyles'
-import {SingleStatLayerConfig} from '../types'
-import {formatStatValue} from '../utils/formatStatValue'
+} from '../../constants/singleStatStyles'
+import {SingleStatLayerConfig} from '../../types'
+import {formatStatValue} from '../../utils/formatStatValue'
 
 interface Props {
   stat: number
