+++ giraffe/src/components/SingleStat/index.ts
@@ -1,0 +1,1 @@
+export {SingleStatLayer} from './SingleStatLayer'
